backend/pyproject.toml
[project]
name = "jamissue-worker"
version = "0.1.0"
description = "JamIssue FastAPI worker pilot"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.6",
  "pydantic-settings==2.7.0",
  "SQLAlchemy==2.0.36",
  "PyMySQL==1.1.1",
  "psycopg[binary]==3.2.9",
  "itsdangerous==2.2.0",
  "cryptography==45.0.7",
  "PyJWT==2.10.1",
  "python-multipart==0.0.20",
]

[dependency-groups]
dev = [
  "workers-py>=0.0.22",
  "workers-runtime-sdk>=0.0.22",
  "pywrangler==0.0.2",
  "pytest==8.3.4",
]